Remove unused youtube cog
