+++ cogs/utils/context.py
@@ -6,7 +6,16 @@ from discord.ext import commands
 
 from config import settings
 
-from ..youtube import to_pages_by_lines
+
+def to_pages_by_lines(content: str, max_size: int):
+    pages = [""]
+    i = 0
+    for line in content.splitlines(keepends=True):
+        if len(pages[i] + line) > max_size:
+            i += 1
+            pages.append("")
+        pages[i] += line
+    return pages
 
 
 def embed_to_string(embed: discord.Embed) -> str:
